# api-go/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API (Go/Fiber). Документация и Try it out: GET /docs/index.html. Для цепочки slug в path используй %2F (например elektronika%2Ftelefony).\n**Socket.IO:** тот же хост и CORS что у REST; путь `/socket.io`; авторизация cookie `session_id`. Неймспейсы `/chat` и `/support`.\nКлиент `socket.io`: надёжнее **v2.x / v3.x**; движок Go (`googollee/go-socket.io` ~v1.7) часто **не коннектится** с **socket.io-client v4** из‑за протокола handshake.\n**/chat** — исходящие: `joinChat` {chatId}, `sendMessage` {chatId, content}, `markAsRead` {chatId}. Входящие: `newMessage`, `messagesRead`, `newChatMessage`.\n**/support** — исходящие: `joinTicket`, `leaveTicket`, `sendSupportMessage` {ticketId, message:{text}}, `supportTyping` {ticketId, isTyping}, `updateTicketStatus` {ticketId, status} (модератор). Входящие: `newSupportMessage`, `moderatorResponse`, `supportUserTyping`, `ticketStatusUpdated`; по неймспейсу `newSupportTicket`; в комнате тикета `ticketAssigned`.",
        "title": "Med Vito REST API (Go)",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:3000",
    "basePath": "/",
    "paths": {
...
        "/review/moderate-review/{id}": {
            "put": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "review-admin"
                ],
                "summary": "Модерация отзыва (admin)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Review id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "APPROVED|DENIDED",
                        "name": "status",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/review/send-review": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "review"
                ],
                "summary": "Оставить отзыв продавцу (сессия)",
                "parameters": [
                    {
                        "description": "Тело",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.swaggerSendReview"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/review/user-reviews/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "review"
                ],
                "summary": "Одобренные отзывы о пользователе",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "User id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
...
        "/subcategory-type/find-by-id/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "subcategory-type"
                ],
                "summary": "Тип подкатегории по id",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
...
        "/subcategory/find-by-id/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "subcategory"
                ],
                "summary": "Подкатегория по id",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
...
        "/type-field/find-by-id/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "type-field"
                ],
                "summary": "Характеристика по id",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
...
    },
    "definitions": {
...
        "main.swaggerSendReview": {
            "type": "object",
            "properties": {
                "rating": {
                    "type": "number",
                    "example": 5
                },
                "reviewedUserId": {
                    "type": "integer"
                },
                "text": {
                    "type": "string"
                }
            }
        },
...
    },
...
}